moved Dtype to mod dtype
