+++ src/lib.rs
@@ -1,5 +1,6 @@
 #![deny(unused_unsafe)]
 
+pub mod dtype;
 pub mod array;
 pub mod device;
 mod utils;
